pyproject.toml
[project]
name = "smollm-trainer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.13.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "tensorflow>=2.21.0",
    "torch>=2.10.0",
    "transformers>=5.3.0",
]
